# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend pendataan perusahaan berbasis kuesioner SE2026 Kuesioner L (UB \u0026 UMKM Bangunan Usaha)",
        "title": "Backend pendataan perusahaan SE2026-L.UB",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/prelist": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Prelist"
                ],
                "summary": "List semua perusahaan target",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Perusahaan"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Prelist"
                ],
                "summary": "Tambah perusahaan target baru",
                "parameters": [
                    {
                        "description": "Perusahaan data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.CreatePrelistRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Perusahaan"
                        }
                    }
                }
            }
        },
        "/prelist/bulk-token": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Prelist"
                ],
                "summary": "Bulk generate token",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/prelist/import": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "Prelist"
                ],
                "summary": "Import perusahaan dari file Excel",
                "parameters": [
                    {
                        "type": "file",
                        "description": "File .xlsx atau .xls",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/prelist/import/template": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Prelist"
                ],
                "summary": "Download template Excel untuk import prelist",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
        "/prelist/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Prelist"
                ],
                "summary": "Detail perusahaan",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Perusahaan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Perusahaan"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Prelist"
                ],
                "summary": "Update data perusahaan",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Perusahaan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Perusahaan data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.UpdatePrelistRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Perusahaan"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Prelist"
                ],
                "summary": "Hapus perusahaan",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Perusahaan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Perusahaan"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/prelist/{id}/token": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Prelist"
                ],
                "summary": "Generate token akses pendataan",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Perusahaan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/wilayah/desa": {
            "get": {
                "tags": [
                    "Wilayah"
                ],
                "summary": "List desa berdasarkan kecamatan",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Kecamatan",
                        "name": "kecamatanId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Desa"
                            }
                        }
                    }
                }
            }
        },
        "/wilayah/kabupaten-kota": {
            "get": {
                "tags": [
                    "Wilayah"
                ],
                "summary": "List kabupaten/kota berdasarkan provinsi",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Provinsi",
                        "name": "provinsiId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.KabupatenKota"
                            }
                        }
                    }
                }
            }
        },
        "/wilayah/kecamatan": {
            "get": {
                "tags": [
                    "Wilayah"
                ],
                "summary": "List kecamatan berdasarkan kabupaten/kota",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Kabupaten/Kota",
                        "name": "kabupatenKotaId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Kecamatan"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.CreatePrelistRequest": {
            "type": "object",
            "required": [
                "nama"
            ],
            "properties": {
                "alamat": {
                    "type": "string",
                    "example": "Jl. Sudirman No. 10, Medan"
                },
                "b1r1": {
                    "type": "string",
                    "example": "12"
                },
                "b1r2": {
                    "type": "string",
                    "example": "1271"
                },
                "b1r3": {
                    "type": "string",
                    "example": "1271010"
                },
                "b1r4": {
                    "type": "string",
                    "example": "1271010001"
                },
                "contactPerson": {
                    "type": "string",
                    "example": "Budi Santoso"
                },
                "email": {
                    "type": "string",
                    "example": "budi@majubersama.co.id"
                },
                "nama": {
                    "type": "string",
                    "example": "PT. Maju Bersama"
                },
                "phone": {
                    "type": "string",
                    "example": "061-12345678"
                },
                "skalaUsaha": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.SkalaUsaha"
                        }
                    ],
                    "example": "UB"
                }
            }
        },
...
        "controllers.UpdatePrelistRequest": {
            "type": "object",
            "properties": {
                "alamat": {
                    "type": "string"
                },
                "b1r1": {
                    "type": "string"
                },
                "b1r2": {
                    "type": "string"
                },
                "b1r3": {
                    "type": "string"
                },
                "b1r4": {
                    "type": "string"
                },
                "contactPerson": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "skalaUsaha": {
                    "$ref": "#/definitions/models.SkalaUsaha"
                }
            }
        },
...
        "models.Desa": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "kecamatanId": {
                    "type": "integer"
                },
                "klas": {
                    "type": "integer"
                },
                "kode": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                }
            }
        },
        "models.KabupatenKota": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "kode": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "provinsiId": {
                    "type": "integer"
                }
            }
        },
        "models.Kecamatan": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "kabupatenKotaId": {
                    "type": "integer"
                },
                "kode": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                }
            }
        },
...
        "models.Perusahaan": {
            "type": "object",
            "properties": {
                "alamat": {
                    "type": "string"
                },
                "b1r1": {
                    "type": "string"
                },
                "b1r2": {
                    "type": "string"
                },
                "b1r3": {
                    "type": "string"
                },
                "b1r4": {
                    "type": "string"
                },
                "contactPerson": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nama": {
                    "type": "string"
                },
                "pendataan": {
                    "$ref": "#/definitions/models.Pendataan"
                },
                "pendataanUmkm": {
                    "$ref": "#/definitions/models.PendataanUmkm"
                },
                "phone": {
                    "type": "string"
                },
                "skalaUsaha": {
                    "$ref": "#/definitions/models.SkalaUsaha"
                },
                "status": {
                    "$ref": "#/definitions/models.StatusPendataan"
                },
                "token": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "models.SkalaUsaha": {
            "type": "string",
            "enum": [
                "UB",
                "UMKM"
            ],
            "x-enum-varnames": [
                "SkalaUB",
                "SkalaUMKM"
            ]
        },
        "models.StatusPendataan": {
            "type": "string",
            "enum": [
                "BELUM",
                "SELESAI"
            ],
            "x-enum-varnames": [
                "StatusBelum",
                "StatusSelesai"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter the token with the `Bearer ` prefix, e.g. \"Bearer abcde12345\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}